CCT becomes instance, not subclass of an algebra.
This gets rid of the metaclass magic. It kind of hurts my heart to do
that, because I was quite fond of the absence of dot notation for
function type definitions, but this is probably more intuitive for
anyone else reading the code, and it is certainly simpler. Moreover, it
opens the way for defining an application operator on definitions later,
so that we could perhaps define functions in terms of other functions
even on the algebra level.