Always bind variables, no free variables.
We now allow signatures that have a callable component. The arguments of
this call are type variables, which will get assigned every time an
instance is requested. This solves the issues caused by free variables:
we can get rid of the `Variables` data structure. Every time a variable
is used, it is also guaranteed to be fresh nw.