Simplify algebra creation.
Instead of defining types on a `TransformationAlgebra` object that are
magically transformed to `Signature`s that know their own name, we will
now just define `Signatures` directly, and make a TransformationAlgebra
(if needed) by passing a dictionary. This doesn't increase work, since
we can still create that dictionary automatically, from `Signature`s
defined in a module:

    algebra = TransformationAlgebra({
        k: v for k, v in dict(globals()).items() if isinstance(v, Σ)
    })

Except now it is explicit what happens.