Attach constraint to variables: one type instance.
We used to keep track of constraints on the top level, and had a
separate `Plain` type instance for constraint-less types. This was
conceptually easy to keep track of, but it also introduced an artificial
need to descend and ascend the objects.

This put the burden of tracking what level we were at on the user of the
library. It also made it harder to implement things like equality.

We now attach constraints directly to variables. This makes the code
considerably simpler.