Simplify arcane schematic type combinator.
Removed the abstract, magical function that combined schematic types. It
hurts a bit to do so, since it was hard to come up with --- but I see
now that it really isn't necessary.

It used to be used for combining schematic types without losing the
property of being schematic --- e.g. you could define a schematic type f
with schematic variables x and y and add a function arrow and the result
would still have schematic variables. However, the only time we *need* a
schematic type is in definitions. Once a definition is instantiated,
it's type *instances* all the way.