Recalculating, not replacing args in abstractions.
Wen substituting a primitive, the resulting type may be more general
than the original type. This was likely the cause for the remaining
errors. Now, abstractions are partial Python functions unless we
really need the abstraction in the final expression tree. This is both
faster and does not cause issues, because we no longer need to replace
variables (whose constraints were probably already tied down) with
(possibly more general) expressions.