Fix bug where nested operations got no input.
We add `ta:feeds` connections from every outer internal node to inner
internal nodes, so that parameter values in the outer operation get
passed along to the beginning of the pipeline of the inner operations.
This solves issue #37, where operations-as-parameters were made aware of
parameter values of the enclosing function, but not if they were nested
inside other operations-as-parameters.