Change structure of `Flow`s and add type checks.
A better DSL was necessary for `Flow`s (workflow queries), because the
previous approach of abusing Python operators was confusing and not
future-proof, as mentioned in issue #39. This commit improves the
situation: we may now use `Flow.serial()`, with tuples for serial flows
and lists for parallel flows.

Additionally, the use of the ellipsis and the necessary recursive types
should now be typechecked.