Remove `Data Definition`s; add `Source Expr`.
The distinction between `Operation`s and `Data` input functions was
nothing but confusing, it necessitated the confusing `Definition`
superclass, and it stood in the way of a more sensible way to
label source nodes. See issues #33 and #51.

It has been done away with. There is now a new `Source` expression that
takes on the function of `Base` expressions with `Data` definitions, and
an `OperationInstance` expression that takes on the function of
`Base` expressions with `Operation` definitions. For now, we do allow
non-function operations: they simply instantiate to `Source`s, so that
our previous functions still work. They will be deprecated eventually.