Build `Language` explicitly from scoped variables.
We have always used a hacky way to add `Operator`s and `Type`s to a
`Language` with as little repetition as possible: we simply passed
`**globals()` to its initialization, and with some magic anything
eligible was added to the language.

The method using `**kwargs` has been removed and there is now an
`add_scope` method that is more explicit about what's actually
happening. This makes me comfortable using it in tests.