Cache `Expr` nodes in `TransformationGraph`s.
This simplifies the code so as to make it easier to generate errors that
occur during workflow concatenation (see issue #30), it makes it easier
to avoid duplication when primitives are expanded later (see issue #40),
and it may make it easier to fix the delayed unification issue (see #31)
by making it possible to construct a full expression for the workflow
without losing node reuse.

I say 'may' because caching every `Expr` seems to significantly slow
down the process; if we are to make choices as to which `Expr` to cache,
we can't be as agnostic.