Constraints use << `__lshift__` not @ `__matmul__`.
At one point (specifically f28de20a21f4d8142de020c9a82982126ab97854) I
decided to use `@` for indicating the constraint that 'this type must
always remain a subtype of the following types'.

The reason was that I wanted to use `<<` and `>>` for subtype bounds
('this type *is* a subtype of the following type'). Subtype *bounds* are
subtly different from the subtype constraints I mentioned earlier in
that the former are fixed by `.resolve()`.

Now, I know that we will never use manual subtype constraints. `<<` is
still a visually intuitive symbol for subtype constraints. It also opens
up `>>` if we ever need to constrain supertypes. More importantly, it
opens up `@` if we want to use it for issue #33.