Begin rewrite of `query.py`.
There were many things wrong with `query.py`.

For example, a unit may be used multiple times, so simply assigning a
variable to a unit will lead to problems: one unit may have multiple
variables, and multiple variables may refer to the same unit.

Adding skips to units poses a problem for the same reason: if a unit is
used once in a series WITH a skip, all subsequent uses of that unit
would also have that skip.

The way it was laid out also made it so that non-units could only appear
in the last place of a series, since there was no way to connect the
variables from the end of a flow to subsequent items.

There are still problems with the approach I'm starting with here ---
it's a first draft.