Rewrite `query.py` further. Allow `OR(Operation)`.
This reintroduces the limitation again that a sequence must consist of
only units except in the last place. I had been working toward removing
that limitation and generalizing the logic (particularly
46957e7b0b276067612e96950858dcfb569020c7 and
00d1ebca0bc674130736a880a37db012397401a1), but that revealed to require
much more complicated logic, and I don't think it's worth wasting too
much time on as we can always rewrite a flow to branch only in the tail
end of sequences. However, we *do* need to be able to specify unions of
operators, as in `OR(f, g)`. That *is* possibly as of this commit.

The tests introduced in 0e7edc41634c8ef694865da91787b83a1b1a5b6d will
need to be rewritten too.