Unify source types with their attachments.
Before, unification happened *during* concatenation and translation to
RDF. This was not a problem when there are no variables in output types,
but there might be cases where variables are already encoded in the RDF
and then unified later. This commit adds a step wherein the type of an
attachment is made a subtype of the type of the point at which it is
attached, *prior* to RDF generation.

This should fix most of issue #31 (though the test does not pass yet).
See also issue #66.