Fix mistake in passthrough.
See issue #81. I didn't write proper tests for it yet as it took an
embarrassingly long time to get it to work in the first place. So, for
reference, the problem was that not every extra source node got
connected to the proper tool transformation output node from which it
was derived, since I forgot that there could be multiple source nodes
referencing the same transformation output. The source nodes really need
to be different nodes, since the whole point of disabling passthrough is
to not communicate type information across tools.