Optionally add Top and Bottom types to taxonomy.
See issue #94. Untested.

This commit changes behaviour, because all sub- and supertypes in a
lineage must now be canonical --- it is no longer possible to have a
canonical root that has non-canonical supertypes.

Also fixed an issue in `.uri()` that goes along with it,
because it allowed unnamed operators to sometimes not throw an error
when put into the taxonomy.