`concretize` accepts more variables than wildcard.
The `TypeInstance.concretize()` method used to accept only wildcards for
replacement. However, this led to issues: parsing `_` as a wildcard
(rather than simply a fresh variable) means that it gets the same
behaviour as wildcards elsewhere: that is, matching against it always
succeeds, which means that we detect recursive types where we shouldn't.
I think there should be a way to deal with this, but it needs some
thought and it's not worth the time: we might as well just parse `_` as
a fresh variable and concretize all variables to the `Top` type. It has
the same effect, since concretization is an explicit step and the user
is not likely to do it without realizing that it will change the meaning
of variables in a type.