Add tests for super/subtype.
Languages will know their full super-/subtype in the future. This is
because in #102, it was decided that, rather than connecting types to
their global taxonomy, the full supertype subtree may be included with
each transformation graph. For this, it is necessary to know the exact
supertypes that are necessary to include.