Simplify successors.
The test for whether a universe was given, to decide whether it makes
sense to go down from the top or up from the bottom, was only performed
for the `Top` and `Bottom` types themselves, but not for types of the
form `F(Top, ...)`. Why? That's why I removed it entirely. Another
option would be to check it always and pass an empty universe in cases
where that would lead to desired behaviour (ie when constructing
canonical types).