Include union-of-types in query.
Until #79 is implemented, as of 11bc3fc7ca61918818f0760db539084dd71c3c6a
we work around the issue of intersection types by using choice, as
described in #77. That is, if we want to query for `R(x, y)`, we also
accept `R(x, y & z)` (or, temporarily using product types for
intersection types, `R(x, y * z)`. However, this means that when we are
querying for workflows that contain an intersection type, we can't just
ignore it: we must query for `{... R(x, y * z)} UNION { ... R(x, y)}`.
This is implemented with this commit, but it should surely not be the
solution we end up with.