Parse `_` as `Top` type directly.
In the previous commit, 685860c0507c1264b354722374966c1b97c549c2, I was
confused as to why exactly I don't parse wildcards as such anymore. I
remember now: it causes recursive types, because the presence of
`wildcard=True` causes the types to always match. It's a solution to
just parse `_` as `Top` directly: the parser should not be able to parse
variables, anyway (... at least now, and I don't see that changing
anytime soon).

However, this sometimes now leads to the following error:

    Could not satisfy subtype `Top` <= `Top`

... which I should address in isolation.