Consistently use `:from` in transformation graphs.
We used to use `ta:from` in transformation queries, but `ta:to` in
transformation graphs. Since they do the same thing, it's better to keep
it consistent and avoid surprises.

`ta:to` is a bit easier to read, because you almost see the arrow going
when you say it. Still, `ta:from` is superior, because usually there
will be more inputs than outputs, and so the Turtle description will be
more brief. That is the reason for settling on `ta:from`.