Avoid edge case where TypeSchema printing fails.
This popped up in <https://github.com/quangis/ratios>, where a wildcard
was part of a subtype constraint, such that the wildcard immediately got
bound, and therefore printing would fail because it was no longer a
wildcard, but it wasn't a named variable either. I don't think it makes
sense to write a test for this right now.